requirements.txt
fastapi~=0.115.12
dishka~=1.6.0
pydantic~=2.11.5
SQLAlchemy~=2.0.41
faststream~=0.5.42
python-dotenv~=1.1.0
pydantic-settings~=2.9.1
aiobotocore~=2.22.0
aiohttp~=3.12.4
requests~=2.32.3
Markdown~=3.8
bs4~=0.0.2
beautifulsoup4~=4.13.4
python-docx~=1.1.2
alembic~=1.16.1
pydub~=0.25.1